refactor: consolidate email endpoints into frappe/email/api.py
Move the whitelisted email endpoints - make (communication composer),
mark_email_as_seen (read-receipt tracking pixel), undo_email_send
(communication/email.py), the inbox actions (email/inbox.py),
get_contact_list / relink / get_communication_doctype (email/__init__.py)
and retry_sending / send_now / toggle_sending (Email Queue doctype) -
into frappe/email/api.py, marked @frappe.public(group="Email") with full
annotations and docstrings. Renames: make -> make_communication, relink
-> relink_communication.

Old dotted paths keep working via lazy module __getattr__ aliases - in
particular the tracking-pixel URL baked into already-sent emails; newly
sent emails embed the canonical path. In-repo JS callers and imports now
use the canonical paths.
