+++ frappe/contacts/doctype/contact/test_contact.py
@@ -2,7 +2,7 @@
 # License: MIT. See LICENSE
 import frappe
 from frappe.contacts.doctype.contact.contact import get_full_name
-from frappe.email import get_contact_list
+from frappe.email.api import get_contact_list
 from frappe.tests import IntegrationTestCase
 
 EXTRA_TEST_RECORD_DEPENDENCIES = ["Contact", "Salutation"]
