+++ frappe/core/doctype/communication/communication.js
@@ -143,7 +143,7 @@ frappe.ui.form.on("Communication", {
 					label: __("Reference Doctype"),
 					fieldname: "reference_doctype",
 					get_query: function () {
-						return { query: "frappe.email.get_communication_doctype" };
+						return { query: "frappe.email.api.get_communication_doctype" };
 					},
 				},
 				{
@@ -169,7 +169,7 @@ frappe.ui.form.on("Communication", {
 					function () {
 						d.hide();
 						frappe.call({
-							method: "frappe.email.relink",
+							method: "frappe.email.api.relink_communication",
 							args: {
 								name: frm.doc.name,
 								reference_doctype: values["reference_doctype"],
@@ -215,7 +215,7 @@ frappe.ui.form.on("Communication", {
 			primary_action_label: __("Move"),
 			primary_action(values) {
 				return frappe.call({
-					method: "frappe.email.inbox.move_email",
+					method: "frappe.email.api.move_email",
 					args: {
 						communication: frm.doc.name,
 						email_account: values.email_account,
@@ -236,7 +236,7 @@ frappe.ui.form.on("Communication", {
 		var flag = "(\\SEEN)";
 
 		return frappe.call({
-			method: "frappe.email.inbox.create_email_flag_queue",
+			method: "frappe.email.api.create_email_flag_queue",
 			args: {
 				names: [frm.doc.name],
 				action: action,
@@ -253,7 +253,7 @@ frappe.ui.form.on("Communication", {
 		var status = frm.doc.status == "Open" ? "Closed" : "Open";
 
 		return frappe.call({
-			method: "frappe.email.inbox.mark_as_closed_open",
+			method: "frappe.email.api.mark_as_closed_open",
 			args: {
 				communication: frm.doc.name,
 				status: status,
@@ -335,7 +335,7 @@ frappe.ui.form.on("Communication", {
 
 	mark_as_spam: function (frm) {
 		frappe.call({
-			method: "frappe.email.inbox.mark_as_spam",
+			method: "frappe.email.api.mark_as_spam",
 			args: {
 				communication: frm.doc.name,
 				sender: frm.doc.sender,
@@ -349,7 +349,7 @@ frappe.ui.form.on("Communication", {
 
 	move_to_trash: function (frm) {
 		frappe.call({
-			method: "frappe.email.inbox.mark_as_trash",
+			method: "frappe.email.api.mark_as_trash",
 			args: {
 				communication: frm.doc.name,
 			},
