+++ frappe/email/__init__.py
@@ -15,41 +15,6 @@ def sendmail_to_system_managers(subject, content):
 	frappe.sendmail(recipients=get_system_managers(), subject=subject, content=content)
 
 
-@frappe.whitelist()
-def get_contact_list(txt: str, page_length: int = 20, extra_filters: str | None = None) -> list[dict]:
-	"""Return email ids for a multiselect field."""
-	if extra_filters:
-		extra_filters = frappe.parse_json(extra_filters)
-
-	filters = [
-		["Contact Email", "email_id", "is", "set"],
-	]
-	if extra_filters:
-		filters.extend(extra_filters)
-
-	fields = ["first_name", "middle_name", "last_name", "company_name"]
-	contacts = frappe.get_list(
-		"Contact",
-		fields=["full_name", "`tabContact Email`.email_id"],
-		filters=filters,
-		or_filters=[[field, "like", f"%{txt}%"] for field in fields]
-		+ [["Contact Email", "email_id", "like", f"%{txt}%"]],
-		limit_page_length=page_length,
-	)
-
-	# The multiselect field will store the `label` as the selected value.
-	# The `value` is just used as a unique key to distinguish between the options.
-	# https://github.com/frappe/frappe/blob/6c6a89bcdd9454060a1333e23b855d0505c9ebc2/frappe/public/js/frappe/form/controls/autocomplete.js#L29-L35
-	return [
-		frappe._dict(
-			value=d.email_id,
-			label=d.email_id,
-			description=d.full_name,
-		)
-		for d in contacts
-	]
-
-
 def get_system_managers():
 	return frappe.db.sql_list(
 		"""select parent FROM `tabHas Role`
@@ -57,60 +22,6 @@ def get_system_managers():
 		AND parent!='Administrator'
 		AND parent IN (SELECT email FROM tabUser WHERE enabled=1)"""
 	)
-
-
-@frappe.whitelist()
-def relink(name: str, reference_doctype: str | None = None, reference_name: str | None = None):
-	frappe.has_permission("Communication", "write", name, throw=True)
-	frappe.db.sql(
-		"""update
-			`tabCommunication`
-		set
-			reference_doctype = %s,
-			reference_name = %s,
-			status = "Linked"
-		where
-			communication_type = "Communication" and
-			name = %s""",
-		(reference_doctype, reference_name, name),
-	)
-
-
-@frappe.whitelist()
-@frappe.validate_and_sanitize_search_inputs
-def get_communication_doctype(
-	doctype: str, txt: str, searchfield: str, start: int, page_len: int, filters: str | list | dict
-):
-	user_perms = frappe.utils.user.UserPermissions(frappe.session.user)
-	user_perms.build_permissions()
-	can_read = user_perms.can_read
-	from frappe import _
-	from frappe.modules import load_doctype_module
-
-	com_doctypes = []
-	if len(txt) < 2:
-		for name in frappe.get_hooks("communication_doctypes"):
-			try:
-				module = load_doctype_module(name, suffix="_dashboard")
-				if hasattr(module, "get_data"):
-					for i in module.get_data()["transactions"]:
-						com_doctypes += i["items"]
-			except ImportError:
-				pass
-	else:
-		com_doctypes = [
-			d[0] for d in frappe.db.get_values("DocType", {"issingle": 0, "istable": 0, "hide_toolbar": 0})
-		]
-
-	results = []
-	txt_lower = txt.lower().replace("%", "")
-
-	for dt in list(set(com_doctypes)):
-		if dt in can_read:
-			if txt_lower in dt.lower() or txt_lower in _(dt).lower():
-				results.append([dt])
-
-	return results
 
 
 def sendmail(
@@ -257,3 +168,21 @@ def sendmail(
 	if now and q:
 		frappe.db.after_commit.add(q.send)
 	return q
+
+
+# `get_contact_list`, `relink`, `get_communication_doctype` moved to frappe.email.api.
+# The aliases keep the old dotted paths working; resolved lazily to avoid
+# circular imports.
+_MOVED_TO_EMAIL_API = {
+	"get_contact_list": "get_contact_list",
+	"relink": "relink_communication",
+	"get_communication_doctype": "get_communication_doctype",
+}
+
+
+def __getattr__(name: str):
+	if new_name := _MOVED_TO_EMAIL_API.get(name):
+		from frappe.email import api
+
+		return getattr(api, new_name)
+	raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
